refactor : http 요청 메서드 AcceptanceTest에서 관리
