--- src/test/java/wooteco/subway/RequestForm.java
@@ -1,35 +1,0 @@
-package wooteco.subway;
-
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import org.springframework.http.MediaType;
-
-public class RequestForm {
-
-    public static ExtractableResponse<Response> createRequest(String path, Object body) {
-        return RestAssured.given().log().all()
-                .body(body)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when()
-                .post(path)
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> findAllRequest(String path) {
-        return RestAssured.given().log().all()
-                .when()
-                .get(path)
-                .then().log().all()
-                .extract();
-    }
-
-    public static ExtractableResponse<Response> deleteRequest(String uri) {
-        return RestAssured.given().log().all()
-                .when()
-                .delete(uri)
-                .then().log().all()
-                .extract();
-    }
-}
