+++ src/test/java/wooteco/subway/section/SectionAcceptanceTest.java
@@ -11,9 +11,6 @@ import wooteco.subway.section.dto.request.SectionCreateRequest;
 import wooteco.subway.station.dto.StationRequest;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static wooteco.subway.RequestForm.createRequest;
-import static wooteco.subway.RequestForm.deleteRequest;
-import static wooteco.subway.line.LineRequestForm.findByIdRequest;
 
 @DisplayName("지하철 구간 관련 기능")
 public class SectionAcceptanceTest extends AcceptanceTest {
@@ -204,7 +201,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         createRequest("/lines/" + 분당선Id + "/sections", 뚝섬_왕십리);
         String uri = "/lines/" + 분당선Id + "/sections?stationId=" + 왕십리역Id;
         ExtractableResponse<Response> response = deleteRequest(uri);
-        ExtractableResponse<Response> 분당선역조회 = findByIdRequest(Long.toString(분당선Id));
+        ExtractableResponse<Response> 분당선역조회 = findByLineIdRequest(Long.toString(분당선Id));
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
@@ -236,7 +233,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         createRequest("/lines/" + 분당선Id + "/sections", 뚝섬_왕십리);
         String uri = "/lines/" + 분당선Id + "/sections?stationId=" + 강남역Id;
         ExtractableResponse<Response> response = deleteRequest(uri);
-        ExtractableResponse<Response> 분당선역조회 = findByIdRequest(Long.toString(분당선Id));
+        ExtractableResponse<Response> 분당선역조회 = findByLineIdRequest(Long.toString(분당선Id));
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
@@ -268,7 +265,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         createRequest("/lines/" + 분당선Id + "/sections", 뚝섬_왕십리);
         String uri = "/lines/" + 분당선Id + "/sections?stationId=" + 잠실역Id;
         ExtractableResponse<Response> response = deleteRequest(uri);
-        ExtractableResponse<Response> 분당선역조회 = findByIdRequest(Long.toString(분당선Id));
+        ExtractableResponse<Response> 분당선역조회 = findByLineIdRequest(Long.toString(분당선Id));
 
         // then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
