--- src/test/java/wooteco/subway/section/SectionRequestForm.java
@@ -1,9 +1,0 @@
-package wooteco.subway.section;
-
-import wooteco.subway.RequestForm;
-
-public class SectionRequestForm extends RequestForm {
-
-    private SectionRequestForm() {
-    }
-}
