--- src/test/java/wooteco/subway/station/StationRequestForm.java
@@ -1,9 +1,0 @@
-package wooteco.subway.station;
-
-import wooteco.subway.RequestForm;
-
-public class StationRequestForm extends RequestForm {
-
-    private StationRequestForm() {
-    }
-}
